# Improve speed of Nearby Look-ups in GridSpaces