--- a/src/spaces/grid_multi.jl
+++ b/src/spaces/grid_multi.jl
@@ -138,9 +138,8 @@ combine_positions(pos, origin, ::GridSpaceIdIterator{false}, ::Bool) = pos .+ or
 function combine_positions(pos, origin, iter::GridSpaceIdIterator{true}, nocheck)
     # the mod function is not needed for many positions and it's expensive compared
     # with checking for bounds so it is better to apply it only as a fallback.
-    off_pos = pos .+ origin
-    (nocheck || checkbounds(Bool, iter.stored_ids, off_pos...)) && return off_pos
-    return mod1.(off_pos, iter.space_size)
+    npos = pos .+ origin
+    (nocheck || checkbounds(Bool, iter.stored_ids, npos...)) ? npos : mod1.(npos, iter.space_size)
 end
 
 # Initialize iteration