--- a/src/spaces/grid_single.jl
+++ b/src/spaces/grid_single.jl
@@ -86,10 +86,15 @@ function nearby_ids(pos::NTuple{D, Int}, model::ABM{<:GridSpaceSingle{D,true}}, 
     nindices = get_offset_indices(model, r)
     stored_ids = model.space.stored_ids
     space_size = size(stored_ids)
-    array_accesses_iterator = (stored_ids[(mod1.(pos .+ β, space_size))...] for β in nindices)
-    # Notice that not all positions are valid; some are empty! Need to filter:
-    valid_pos_iterator = Base.Iterators.filter(x -> x ≠ 0, array_accesses_iterator)
-    return valid_pos_iterator
+    position_iterator = (pos .+ β for β in nindices)
+    if all(i -> r < pos[i] <= space_size[i] - r, 1:D)
+        ids_iterator = (stored_ids[p...] for p in position_iterator 
+                        if stored_ids[p...] != 0)
+    else
+        ids_iterator = (stored_ids[mod1.(p, space_size)...] for p in position_iterator
+                        if stored_ids[mod1.(p, space_size)...] != 0)
+    end
+    return ids_iterator
 end
 
 function nearby_ids(pos::NTuple{D, Int}, model::ABM{<:GridSpaceSingle{D,false}}, r = 1,
@@ -97,16 +102,18 @@ function nearby_ids(pos::NTuple{D, Int}, model::ABM{<:GridSpaceSingle{D,false}},
     ) where {D}
     nindices = get_offset_indices(model, r)
     stored_ids = model.space.stored_ids
-    positions_iterator = (pos .+ β for β in nindices)
-    # Here we combine in one filtering step both valid accesses to the space array
-    # but also that the accessed location is not empty (i.e., id is not 0)
-    array_accesses_iterator = Base.Iterators.filter(
-        pos -> checkbounds(Bool, stored_ids, pos...) && stored_ids[pos...] ≠ 0,
-        positions_iterator
-    )
-    return (stored_ids[pos...] for pos in array_accesses_iterator)
+    space_size = size(stored_ids)
+    position_iterator = (pos .+ β for β in nindices)
+    if all(i -> r < pos[i] <= space_size[i] - r, 1:D)
+        ids_iterator = (stored_ids[p...] for p in position_iterator 
+                        if stored_ids[p...] != 0)
+    else
+        ids_iterator = (stored_ids[p...] for p in position_iterator 
+                        if checkbounds(Bool, stored_ids, p...) && stored_ids[p...] != 0)
+    end
+    return ids_iterator
 end
-
+ 
 # Contrary to `GridSpace`, we also extend here `nearby_ids(a::Agent)`.
 # Because, we know that one agent exists per position, and hence we can skip the
 # call to `filter(id -> id ≠ a.id, ...)` that happens in core/space_interaction_API.jl.
@@ -115,4 +122,16 @@ end
 function nearby_ids(
     a::A, model::ABM{<:GridSpaceSingle{D},A}, r = 1) where {D,A<:AbstractAgent}
     return nearby_ids(a.pos, model, r, offsets_within_radius_no_0)
+end
+
+# Method for walking on GridSpaceSingle, similar to `walk!` in spaces/utilities.jl
+function walk!(
+    agent::AbstractAgent,
+    direction::NTuple{D,Int},
+    model::ABM{<:GridSpaceSingle}
+) where {D}
+    target = normalize_position(agent.pos .+ direction, model)
+    if isempty(target, model) # if target unoccupied
+        move_agent!(agent, target, model)
+    end
 end