This came to my mind today, it's a simple change but has a big impact: 
Flocking in the large version is 1.5x faster now! edit: but with a big catch :D, see below

In general it has a good impact when periodic is true and just a slight impact when it is false.

It's time for 5.15 I guess  :D