@Tortar 2023-06-03
Wait a second, I don't see much of the speed-up I saw earlier, need to investigate a bit better