@Tortar 2023-06-03
Now we can actually evaluate the matter:

```julia
using Agents, BenchmarkTools

@agent A GridAgent{2} begin
end

function periodic_model(GridType; numagents = 1000, griddims = (50, 50))
    space = GridType(griddims, periodic = true)
    model = ABM(A, space)
    for _ in 1:numagents
        add_agent_single!(model)
    end
    return model
end

function aperiodic_model(GridType; numagents = 1000, griddims = (50, 50))
    space = GridType(griddims, periodic = false)
    model = ABM(A, space)
    for _ in 1:numagents
        add_agent_single!(model)
    end
    return model
end

pos = (20,20)

# ids in gridspace
@btime collect(nearby_ids($pos, model, 1)) setup=(model=periodic_model(GridSpace))
@btime collect(nearby_ids($pos, model, 1)) setup=(model=aperiodic_model(GridSpace))
@btime collect(nearby_ids($pos, model, 10)) setup=(model=periodic_model(GridSpace))
@btime collect(nearby_ids($pos, model, 10)) setup=(model=aperiodic_model(GridSpace))

# ids in gridspacesingle
@btime collect(nearby_ids($pos, model, 1)) setup=(model=periodic_model(GridSpaceSingle))
@btime collect(nearby_ids($pos, model, 1)) setup=(model=aperiodic_model(GridSpaceSingle))
@btime collect(nearby_ids($pos, model, 10)) setup=(model=periodic_model(GridSpaceSingle))
@btime collect(nearby_ids($pos, model, 10)) setup=(model=aperiodic_model(GridSpaceSingle))

# positions
@btime collect(nearby_positions($pos, model, 1)) setup=(model=periodic_model(GridSpace))
@btime collect(nearby_positions($pos, model, 1)) setup=(model=aperiodic_model(GridSpace))
@btime collect(nearby_positions($pos, model, 10)) setup=(model=periodic_model(GridSpace))
@btime collect(nearby_positions($pos, model, 10)) setup=(model=aperiodic_model(GridSpace))
```

gives before and after this PR:

```julia
# before PR

julia> # ids in gridspace
       @btime collect(nearby_ids($pos, model, 1)) setup=(model=periodic_model(GridSpace))
  59.522 ns (1 allocation: 64 bytes)

julia> @btime collect(nearby_ids($pos, model, 1)) setup=(model=aperiodic_model(GridSpace))
  42.738 ns (1 allocation: 64 bytes)

julia> @btime collect(nearby_ids($pos, model, 10)) setup=(model=periodic_model(GridSpace))
  4.242 μs (150 allocations: 9.66 KiB)

julia> @btime collect(nearby_ids($pos, model, 10)) setup=(model=aperiodic_model(GridSpace))
  1.843 μs (5 allocations: 2.67 KiB)

julia> # ids in gridspacesingle
       @btime collect(nearby_ids($pos, model, 1)) setup=(model=periodic_model(GridSpaceSingle))
  71.976 ns (1 allocation: 64 bytes)

julia> @btime collect(nearby_ids($pos, model, 1)) setup=(model=aperiodic_model(GridSpaceSingle))
  40.997 ns (1 allocation: 64 bytes)

julia> @btime collect(nearby_ids($pos, model, 10)) setup=(model=periodic_model(GridSpaceSingle))
  3.830 μs (6 allocations: 2.92 KiB)

julia> @btime collect(nearby_ids($pos, model, 10)) setup=(model=aperiodic_model(GridSpaceSingle))
  1.815 μs (6 allocations: 2.74 KiB)

julia> # positions
       @btime collect(nearby_positions($pos, model, 1)) setup=(model=periodic_model(GridSpace))
  83.006 ns (1 allocation: 193 bytes)

julia> @btime collect(nearby_positions($pos, model, 1)) setup=(model=aperiodic_model(GridSpace))
  107.129 ns (3 allocations: 273 bytes)

julia> @btime collect(nearby_positions($pos, model, 10)) setup=(model=periodic_model(GridSpace))
  3.320 μs (2 allocations: 8.61 KiB)

julia> @btime collect(nearby_positions($pos, model, 10)) setup=(model=aperiodic_model(GridSpace))
  3.310 μs (7 allocations: 16.67 KiB)

# after PR

julia> # ids in gridspace
       @btime collect(nearby_ids($pos, model, 1)) setup=(model=periodic_model(GridSpace))
  61.147 ns (1 allocation: 64 bytes)

julia> @btime collect(nearby_ids($pos, model, 1)) setup=(model=aperiodic_model(GridSpace))
  41.979 ns (1 allocation: 64 bytes)

julia> @btime collect(nearby_ids($pos, model, 10)) setup=(model=periodic_model(GridSpace))
  4.212 μs (147 allocations: 9.52 KiB)

julia> @btime collect(nearby_ids($pos, model, 10)) setup=(model=aperiodic_model(GridSpace))
  2.881 μs (144 allocations: 9.27 KiB)

julia> # ids in gridspacesingle
       @btime collect(nearby_ids($pos, model, 1)) setup=(model=periodic_model(GridSpaceSingle))
  42.718 ns (2 allocations: 112 bytes)

julia> @btime collect(nearby_ids($pos, model, 1)) setup=(model=aperiodic_model(GridSpaceSingle))
  42.495 ns (2 allocations: 112 bytes)

julia> @btime collect(nearby_ids($pos, model, 10)) setup=(model=periodic_model(GridSpaceSingle))
  1.646 μs (7 allocations: 2.78 KiB)

julia> @btime collect(nearby_ids($pos, model, 10)) setup=(model=aperiodic_model(GridSpaceSingle))
  1.657 μs (7 allocations: 2.78 KiB)

julia> # positions
       @btime collect(nearby_positions($pos, model, 1)) setup=(model=periodic_model(GridSpace))
  37.874 ns (2 allocations: 224 bytes)

julia> @btime collect(nearby_positions($pos, model, 1)) setup=(model=aperiodic_model(GridSpace))
  38.491 ns (2 allocations: 224 bytes)

julia> @btime collect(nearby_positions($pos, model, 10)) setup=(model=periodic_model(GridSpace))
  449.297 ns (2 allocations: 7.11 KiB)

julia> @btime collect(nearby_positions($pos, model, 10)) setup=(model=aperiodic_model(GridSpace))
  414.357 ns (2 allocations: 7.10 KiB)
```

which means that the only case where this optimizations has no effect is actually GridSpace :D , but I think that this is because I did it rather blindly and I started from the other cases, and then I applied it a bit in a hurry and was fooled by wrong benchmarks

BUT it is almost 10x with a radius of 10 for `nearby_positions` !! And 2x in general for `nearby_ids` in a periodic `GridSpaceSingle`