@Tortar 2023-06-03
> ```julia
> (i -> r < pos[i] <= space_size[i] - r, 1:D)
> ```

If I recall correctly the positions starts from 1 and the extent of the space is inclusive so it should be right