@Tortar 2023-06-04
Incredibly enough, even this implementation for `GridSpace` seems to bring some benefit, I ran Flocking multiple times but using Periodic = False and in this case it becomes faster, really almost 2x in the large model, while with periodic = True as it is the correct Flocking everything is still the same speed. Don't know what to think actually, these results are strange, but didn't find any other way to use the idea to optimize the code further. So I think everything should be mergeable as it is now