@Tortar 2023-06-04
> Okay, so if I understand this PR right, it checks if the chebyshev distance of the cell is further away from the walls than the radius. If it is, it skips completely doing the bounds check or the mod1 operation. That makes sense, pretty stupid of us to not do it from the very first implementation.

that's exactly how it works :-)